# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BaZi Master API",
    "version": "2.0.0",
    "description": "八字 / 紫微斗数 / 塔罗 / 周易 / 星座 / 合盘的算法能力层。\n\n**无状态**：不存数据、不认用户、没有数据库。同样的入参永远得到同样的结果，\n可以随意水平扩容，也不需要迁移或备份。\n\n**无鉴权**：所有端点都是公开的，靠限流而不是身份来控制成本。\n需要访问控制请放在反向代理层。唯一带凭据的是 `/api-docs`（Basic）和 `/metrics`（Bearer），\n它们都不是能力端点。\n\n`/ai-interpret` 这类端点会调用外部大模型，是唯一有副作用（花钱、可能超时）的一类；\n排盘类端点纯本地计算。"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "API 服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "metricsToken": {
        "type": "http",
        "scheme": "bearer"
      },
      "docsBasic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误信息"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/fengshui/bazhai": {
      "post": {
        "tags": [
          "风水"
        ],
        "summary": "八宅命卦与八方吉凶",
        "description": "由出生年与性别定本命卦（男 11 减、女加四，得五者男寄坤女寄艮），再以变爻法排八方游年星。给了月日则以**立春**为界定年，不是元旦。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "birthYear",
                  "gender"
                ],
                "properties": {
                  "birthYear": {
                    "type": "integer"
                  },
                  "birthMonth": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 12
                  },
                  "birthDay": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 31
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "male",
                      "female"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "lifeTrigram": {
                      "type": "object",
                      "description": "本命卦、卦数、方位、东西四命"
                    },
                    "younian": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      },
                      "description": "八方游年星"
                    },
                    "auspiciousDirections": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "inauspiciousDirections": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "缺出生年或性别",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/fengshui/almanac": {
      "get": {
        "tags": [
          "择吉"
        ],
        "summary": "当日历注（建除、值宿、吉神凶煞、彭祖百忌）",
        "description": "不给日期则取服务器当日。历注数据来自 lunar-javascript。",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "month",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "day",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ganzhi": {
                      "type": "object"
                    },
                    "lunarDate": {
                      "type": "object"
                    },
                    "zhiXing": {
                      "type": "string",
                      "description": "建除十二神"
                    },
                    "xiu": {
                      "type": "object",
                      "description": "二十八宿及其吉凶"
                    },
                    "auspiciousGods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "inauspiciousGods": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "pengzu": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "日期非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/fengshui/name": {
      "post": {
        "tags": [
          "姓名学"
        ],
        "summary": "姓名五格与三才",
        "description": "**笔画数由调用方提供**，引擎不内置字典 —— 康熙笔画与简体笔画差异很大，部首另有独立算法（如「氵」按「水」计四画），内置一份来路不明的笔画表只会让结果看着精确、实则不可追溯。五格算法本身是确定的。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "surnameStrokes",
                  "givenNameStrokes"
                ],
                "properties": {
                  "surnameStrokes": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "description": "姓的逐字笔画"
                  },
                  "givenNameStrokes": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "minimum": 1
                    },
                    "description": "名的逐字笔画"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "grids": {
                      "type": "object",
                      "description": "天格、人格、地格、外格、总格"
                    },
                    "gridElements": {
                      "type": "object",
                      "description": "各格五行（按个位取）"
                    },
                    "sancai": {
                      "type": "object",
                      "description": "三才配置"
                    },
                    "sancaiRelations": {
                      "type": "object",
                      "description": "天人、人地的生克关系"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "笔画数缺失或非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}